[MODIFY] services 파일 삭제 후 consumers로 통합
